Move ParamUtils out of hash_tir::utils and remove AccessToUtils
