+++ compiler/hash-tir/src/params/mod.rs
@@ -26,6 +26,8 @@ use crate::{
     tuples::TupleTy,
 };
 
+pub mod utils;
+
 /// A parameter, declaring a potentially named variable with a given type and
 /// possibly a default value.
 #[derive(Debug, Clone, Copy)]
