+++ compiler/hash-typecheck/src/unification.rs
@@ -10,11 +10,11 @@ use hash_tir::{
     fns::{CallTerm, FnTy},
     holes::Hole,
     lits::Lit,
-    params::ParamsId,
+    params::{utils::validate_params, ParamsId},
     sub::Sub,
     symbols::SymbolId,
     terms::{Term, TermId, Ty, TyId},
-    utils::{traversing::Atom, AccessToUtils},
+    utils::traversing::Atom,
 };
 use hash_utils::derive_more::Deref;
 use once_cell::unsync::OnceCell;
@@ -374,8 +374,8 @@ impl<'tc, T: AccessToTypechecking> UnificationOps<'tc, T> {
         in_param_scope: impl FnOnce() -> TcResult<U>,
     ) -> TcResult<U> {
         // Validate the parameters and ensure they are of the same length
-        self.param_utils().validate_params(src_id)?;
-        self.param_utils().validate_params(target_id)?;
+        validate_params(src_id)?;
+        validate_params(target_id)?;
         if src_id.len() != target_id.len() {
             return Err(TcError::WrongParamLength {
                 given_params_id: src_id,
